+++ CHANGELOG.md
@@ -30,6 +30,7 @@ _Work in progress._
   validation but before starting the service.
 - Add an `allowed_runners` setting to provide a list of the runners allowed in
   submitted jobs.
+- Add a `generic/gpg_key` identity to store GPG keys.
 
 ### Bug fixes
 - Fix job pagination in evcli.